moved nav bar to own class
